Moving password genrator into its own file.
