+++ scripts/frontend/import_users.py
@@ -3,9 +3,8 @@ import csv
 import logging
 import os
 
-from password_generator import PasswordGenerator
-
 from ga_dashboard.grafana_ga.folder import GrafanaGAFolder
+from ga_dashboard.grafana_ga.password_gen import initialise_strict_password_generator
 from ga_dashboard.grafana_ga.user import GrafanaGAUser
 
 logger = logging.getLogger(__name__)
@@ -31,33 +30,6 @@ Example (using Grafana admin defaults) from `GA4HPCdashboard` top directory:
 python scripts/frontend/import_users.py --input_file ga_dashboard/samples/grafana_users_list.csv --admin_login admin --admin_password admin
 
 """ 
-
-def initialise_strict_password_generator() -> PasswordGenerator:
-    """
-    Sets up a password generator. See https://pypi.org/project/random-password-generator/
-     
-    This produces passwords which adhere to Grafana password policy, if enforced.
-    See https://grafana.com/docs/grafana/next/setup-grafana/configure-security/configure-authentication/grafana/#strong-password-policy
-    """
-    strict_pwo = PasswordGenerator()
-
-    # At least 12 characters
-    strict_pwo.minlen = 12
-    
-    # At least one uppercase letter
-    strict_pwo.minuchars = 1
-
-    # At least one lowercase letter
-    strict_pwo.minlchars = 1
-
-    # At least one number
-    strict_pwo.minnumbers = 1
-
-    # At least one special character
-    strict_pwo.minschars = 1
-
-    return strict_pwo
-
 
 def main():
     argparser = argparse.ArgumentParser(description="Import users in user list file to Grafana.")
