+++ scripts/frontend/run_dashboard.py
@@ -3,9 +3,10 @@ import csv
 import logging
 import os
 
+from ga_dashboard.grafana_ga.dashboard import GrafanaGADashboard
 from ga_dashboard.grafana_ga.datasource import GrafanaGADataSource
 from ga_dashboard.grafana_ga.folder import GrafanaGAFolder
-from ga_dashboard.grafana_ga.dashboard import GrafanaGADashboard
+from ga_dashboard.grafana_ga.password_gen import initialise_strict_password_generator
 from ga_dashboard.grafana_ga.user import GrafanaGAUser
 
 
@@ -28,7 +29,7 @@ Combines the usage of several of the scripts:
 # % python scripts/frontend/run_dashboard.py --admin_password <grafana_admin_password> \
 #        --input_file ga_dashboard/samples/grafana_users_list.csv \
 #        --db_name ga_db --db_user postgres --db_password <db_password> \
-#        --input_dir ga_dashboard/dashboards    <-- If not set, uses default dashboard .json directory.
+#        --input_dir ga_dashboard/dashboards
 #        --name grafana-postgresql-datasource   <-- If not set, uses "grafana-postgresql-ga_db" 
 
 # Or: as above but:
@@ -133,15 +134,23 @@ def main():
     logger.info('###########################')
     grafana_user = GrafanaGAUser(login, password, grafana_url, ga_dashboard_folder_name)
 
+    # Set up password generator
+    PWG = initialise_strict_password_generator()
+
     with open(input_file, encoding='utf-8-sig', newline='') as csvfile:
         reader = csv.DictReader(csvfile, delimiter=',', quotechar='"')
         for row in reader:
             # Create team (if needed)
             grafana_user.create_team(row['Group'])   #grafana_user.create_team(row['Team name'])
 
+            # Generate password. It's up to you to decide what to do with it.
+            grafana_password = PWG.generate()
+            row['GrafanaPassword'] = grafana_password
+
             # Create user (if needed)
             row['org_id'] = 1 # Default organisation
-            grafana_user.create_user(row)
+            if (grafana_user.create_user(row)):
+                logger.info(f"** Grafana password for user {row['User']} is {grafana_password} **")
 
     # Folder
     logger.info('##############################')
